pyproject.toml
[build-system] # Require setuptool version due to https://github.com/pypa/setuptools/issues/2938
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

 # See https://python-poetry.org/docs/pyproject/ for more keywords for the project table
[project]
name = "NNFE"
version = "0.1.0"
description = "NNFE: Implementation of NNFE with CARDIAX as backend"
readme = "README.md"
requires-python = ">=3.9, <3.14" # Needed for TACC, default 3.9.7
license = { text = "GPL-3.0-only" }

authors = [
    {name = "Benjamin Thomas", email = "bjt2369@utexas.edu"}
    ]

# Direct dependencies
dependencies = [
    'jax>=0.4', # JAX is used for automatic differentiation and GPU acceleration
    'numpy>=2.0', # Numpy is used for numerical operations
    'meshio>=5.0', # Meshio is used for reading and writing mesh files
    'lineax>=0.1.0', # Lineax is used for matrix operators on GPUs
    'jaxtyping>=0.3', # Jaxtyping is used for type annotations in JAX
    'fenics-basix>=0.9', # Basix is used for finite element basis functions
    'pyyaml>=6.0', # PyYAML is used for reading configuration files
    'pygmsh>=7.0', # pygmsh is used for mesh generation
    'pyvista>=0.43', # PyVista is used for 3D visualization
    'vtk < 9.6', # VTK is used for 3D visualization
    'optax>=0.2.0' # Optax is used for optimization in JAX
]

[project.optional-dependencies]
# For GPU, calls jax GPU install
cuda13 = ["jax[cuda13]>=0.4"]
cuda13-local = ["jax[cuda13-local]>=0.4"]

test = [
   "ruff", # Formatting: https://docs.astral.sh/ruff/rules/#pydocstyle-d
   "pytest", # Testing suite
   "pytest-cov" # Coverage reports
]

docs = [
    "mkdocs==1.6.1",
    "mkdocs-autorefs==1.4.3",
    "mkdocs-get-deps==0.2.0",
    "mkdocs-material==9.7.1",
    "mkdocs-material-extensions==1.3.1",
    "mkdocstrings==1.0.1",
    "mkdocstrings-python==2.0.1"    
]

# Used for building wheels and uploading to pypi
build = ["build"]

all = ["nnfe[test,docs,build,cuda13]"]

[tool.mypy]
ignore_missing_imports = true # Does not show errors when importing untyped libraries
exclude = [ # We only want mypy to consider files that are not generated in installing or building documentation
    "docs/",
    "build/"
]
files = [  # Folder to which files that should be checked by mypy
    "src",
    "test"
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
    "--cov=nnfe",
    "--cov-report=html",
    "--cov-report=term-missing",
    "-v"]

testpaths = [
    "test"
]


# Pair notebooks with python files
[tool.jupytext]
formats = "demos///ipynb,demos///py:percent"